backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "addict>=2.4.0",
    "datasets>=4.8.4",
    "fastapi>=0.115.0",
    "fastmcp>=2.0.0",
    "ffmpeg-python>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=1.2.22",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=0.2.0",
    "modelscope>=1.35.3",
    "funasr>=1.0.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.8.0",
    "simplejson>=3.20.2",
    "sortedcontainers>=2.4.0",
    "soundfile>=0.13.1",
    "torch>=2.11.0",
    "uvicorn>=0.32.0",
    "torchaudio>=2.11.0",
    "rapidocr-onnxruntime>=1.4.4",
]
